[ADD] suggest (건의페이지) 퍼블리싱 및 컴포넌트 구조 작업 완료
