+++ src/pages/ReserveHome.jsx
@@ -26,15 +26,15 @@ function ReserveHome() {
 
   // getTimezoneOffset()은 현재 시간과의 차이를 분 단위로 반환하는 함수 => 9시간 이므로 -540 이 나올거임
   const offset = new Date().getTimezoneOffset() * 60000; // ms 단위를 맞추기 위해 60000 곱해줌
-  const today = new Date();
-  // console.log(today);
+  const today = new Date(Date.now() - offset);
+  // console.log(today.getDay());
 
   const nowDate = today.toISOString().split('T')[0];
   // console.log(nowDate);
 
   const [selectedDate, setSelectedDate] = useState({
     pickDate: nowDate,
-    pickDay: dayList[new Date().getDay()],
+    pickDay: dayList[today.getDay()],
   });
   const [listDayTable, setListDayTable] = useState([]);
   // console.log(listDayTable);
